Feat. add useGetSingleDatabase
